kata-webhook: simplify building kata-webhook container
Before one can build the container image the vendor directory
from the parent directory has to be copied to the kata-webhook
directory. Otherwise the build will fail because it can't find
all necessary packages. See issue
https://github.com/kata-containers/tests/issues/2479

We can avoid this by copying in the vendor directory during the
docker build. By having docker do it for us we don't need the manual
step any more and can remove it from the README.

The vendor directory is ~28 megs and it takes less than a second
to copy it into the build env, so no noticeable slow down.

We can also get rid of the .gitignore entry for the vendor dir when
it is only copied in for the container build.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>